backend/requirements.txt
Django==2.2.19
pytz==2023.3
sqlparse==0.4.4
psycopg2-binary==2.9.3
python-dotenv==1.0.0
webcolors-1.13
Pillow-10.0.0
djangorestframework-3.14.0
djangorestframework-simplejwt-4.7.2
djoser-2.1.0
django-filter-23.2